refactor: extract preferences/skills/skill-file sections
Move the last three settings sections into per-section components.
